# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mirador RCA Engine REST API",
    "description": "REST API for the Mirador Root Cause Analysis Engine",
    "version": "1.0.0",
    "contact": {
      "name": "Mirador Platform Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://mirador-rca.internal",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/v1/investigate": {
      "post": {
        "summary": "Investigate Incident",
        "description": "Initiates a root cause analysis investigation for an incident",
        "operationId": "investigateIncident",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Investigation completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrelationResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/correlations": {
      "get": {
        "summary": "List Correlations",
        "description": "Lists historical correlations with optional filtering",
        "operationId": "listCorrelations",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Tenant identifier"
          },
          {
            "name": "service",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Service name filter"
          },
          {
            "name": "start",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Start time for correlation window"
          },
          {
            "name": "end",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "End time for correlation window"
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "description": "Number of results per page"
          },
          {
            "name": "page_token",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Token for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "List of correlations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCorrelationsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patterns": {
      "get": {
        "summary": "Get Patterns",
        "description": "Retrieves known failure patterns",
        "operationId": "getPatterns",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Tenant identifier"
          },
          {
            "name": "service",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Service name filter"
          }
        ],
        "responses": {
          "200": {
            "description": "List of failure patterns",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPatternsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/feedback": {
      "post": {
        "summary": "Submit Feedback",
        "description": "Submits user feedback for a correlation result",
        "operationId": "submitFeedback",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feedback"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Feedback submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackAck"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Returns the service health status",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "InvestigationRequest": {
        "type": "object",
        "required": [
          "incident_id",
          "time_range",
          "tenant_id"
        ],
        "properties": {
          "incident_id": {
            "type": "string",
            "description": "Unique identifier for the incident"
          },
          "symptoms": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of symptoms observed"
          },
          "time_range": {
            "$ref": "#/components/schemas/TimeRange"
          },
          "affected_services": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Services affected by the incident"
          },
          "anomaly_threshold": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 1,
            "default": 0.5,
            "description": "Threshold for anomaly detection"
          },
          "tenant_id": {
            "type": "string",
            "description": "Tenant identifier"
          }
        }
      },
      "TimeRange": {
        "type": "object",
        "required": [
          "start",
          "end"
        ],
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "description": "Start time of the analysis window"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "description": "End time of the analysis window"
          }
        }
      },
      "CorrelationResult": {
        "type": "object",
        "required": [
          "correlation_id",
          "incident_id",
          "root_cause",
          "confidence",
          "created_at"
        ],
        "properties": {
          "correlation_id": {
            "type": "string",
            "description": "Unique identifier for the correlation"
          },
          "incident_id": {
            "type": "string",
            "description": "Associated incident identifier"
          },
          "root_cause": {
            "type": "string",
            "description": "Identified root cause"
          },
          "confidence": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 1,
            "description": "Confidence score of the analysis"
          },
          "affected_services": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Services affected by the root cause"
          },
          "red_anchors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RedAnchor"
            },
            "description": "Key anomaly indicators"
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "description": "Chronological event sequence"
          },
          "recommendations": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Suggested remediation actions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the correlation was created"
          }
        }
      },
      "RedAnchor": {
        "type": "object",
        "required": [
          "service",
          "selector",
          "data_type",
          "timestamp",
          "anomaly_score",
          "threshold"
        ],
        "properties": {
          "service": {
            "type": "string",
            "description": "Service name"
          },
          "selector": {
            "type": "string",
            "description": "Metric/log/trace selector"
          },
          "data_type": {
            "$ref": "#/components/schemas/DataType"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "When the anomaly occurred"
          },
          "anomaly_score": {
            "type": "number",
            "format": "float",
            "description": "Anomaly score"
          },
          "threshold": {
            "type": "number",
            "format": "float",
            "description": "Anomaly threshold"
          }
        }
      },
      "TimelineEvent": {
        "type": "object",
        "required": [
          "time",
          "event",
          "service",
          "severity",
          "anomaly_score",
          "data_source"
        ],
        "properties": {
          "time": {
            "type": "string",
            "format": "date-time",
            "description": "Event timestamp"
          },
          "event": {
            "type": "string",
            "description": "Event description"
          },
          "service": {
            "type": "string",
            "description": "Associated service"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "anomaly_score": {
            "type": "number",
            "format": "float",
            "description": "Anomaly score"
          },
          "data_source": {
            "$ref": "#/components/schemas/DataType"
          }
        }
      },
      "DataType": {
        "type": "string",
        "enum": [
          "metrics",
          "logs",
          "traces"
        ],
        "description": "Type of observability data"
      },
      "Severity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "description": "Severity level"
      },
      "ListCorrelationsRequest": {
        "type": "object",
        "properties": {
          "tenant_id": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "page_size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "page_token": {
            "type": "string"
          }
        }
      },
      "ListCorrelationsResponse": {
        "type": "object",
        "required": [
          "correlations"
        ],
        "properties": {
          "correlations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CorrelationResult"
            }
          },
          "next_page_token": {
            "type": "string",
            "description": "Token for next page of results"
          }
        }
      },
      "FailurePattern": {
        "type": "object",
        "required": [
          "id",
          "name",
          "services",
          "anchor_templates",
          "prevalence",
          "precision",
          "recall"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Pattern identifier"
          },
          "name": {
            "type": "string",
            "description": "Pattern name"
          },
          "description": {
            "type": "string",
            "description": "Pattern description"
          },
          "services": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Associated services"
          },
          "anchor_templates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnchorTemplate"
            },
            "description": "Pattern templates"
          },
          "prevalence": {
            "type": "number",
            "format": "float",
            "description": "How often this pattern occurs"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "description": "When this pattern was last observed"
          },
          "precision": {
            "type": "number",
            "format": "float",
            "description": "Pattern precision"
          },
          "recall": {
            "type": "number",
            "format": "float",
            "description": "Pattern recall"
          }
        }
      },
      "AnchorTemplate": {
        "type": "object",
        "required": [
          "service",
          "signal_type",
          "selector",
          "threshold"
        ],
        "properties": {
          "service": {
            "type": "string",
            "description": "Service name"
          },
          "signal_type": {
            "type": "string",
            "description": "Type of signal"
          },
          "selector": {
            "type": "string",
            "description": "Signal selector"
          },
          "typical_lead_lag": {
            "type": "number",
            "format": "float",
            "description": "Typical timing offset"
          },
          "threshold": {
            "type": "number",
            "format": "float",
            "description": "Anomaly threshold"
          }
        }
      },
      "GetPatternsResponse": {
        "type": "object",
        "required": [
          "patterns"
        ],
        "properties": {
          "patterns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FailurePattern"
            }
          }
        }
      },
      "Feedback": {
        "type": "object",
        "required": [
          "tenant_id",
          "correlation_id",
          "correct"
        ],
        "properties": {
          "tenant_id": {
            "type": "string",
            "description": "Tenant identifier"
          },
          "correlation_id": {
            "type": "string",
            "description": "Correlation to provide feedback for"
          },
          "correct": {
            "type": "boolean",
            "description": "Whether the correlation was correct"
          },
          "notes": {
            "type": "string",
            "description": "Additional feedback notes"
          }
        }
      },
      "FeedbackAck": {
        "type": "object",
        "required": [
          "correlation_id",
          "accepted"
        ],
        "properties": {
          "correlation_id": {
            "type": "string",
            "description": "Correlation identifier"
          },
          "accepted": {
            "type": "boolean",
            "description": "Whether feedback was accepted"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "SERVING"
            ],
            "description": "Service health status"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      }
    }
  }
}